[cudaaligner] Use int32_t for RLE

+++ cudaaligner/include/claraparabricks/genomeworks/cudaaligner/alignment.hpp
@@ -92,7 +92,7 @@ public:
     ///         mistmatch and insertions in alignment.
     virtual const std::vector<AlignmentState>& get_alignment() const = 0;
     virtual const std::vector<int8_t>& get_actions() const           = 0;
-    virtual const std::vector<uint8_t>& get_runlengths() const       = 0;
+    virtual const std::vector<int32_t>& get_runlengths() const       = 0;
 
     /// \brief Get the edit distance corrsponding to the alignment
     ///
